+++ core/constraint-streams/src/main/java/ai/timefold/solver/constraint/streams/bavet/bi/ConcatBiUniNode.java
@@ -3,11 +3,12 @@ package ai.timefold.solver.constraint.streams.bavet.bi;
 import ai.timefold.solver.constraint.streams.bavet.common.AbstractConcatNode;
 import ai.timefold.solver.constraint.streams.bavet.common.tuple.BiTuple;
 import ai.timefold.solver.constraint.streams.bavet.common.tuple.TupleLifecycle;
+import ai.timefold.solver.constraint.streams.bavet.common.tuple.UniTuple;
 
-final class BavetBiBiConcatNode<A, B>
-        extends AbstractConcatNode<BiTuple<A, B>, BiTuple<A, B>, BiTuple<A, B>> {
+final class ConcatBiUniNode<A, B>
+        extends AbstractConcatNode<BiTuple<A, B>, UniTuple<A>, BiTuple<A, B>> {
 
-    BavetBiBiConcatNode(TupleLifecycle<BiTuple<A, B>> nextNodesTupleLifecycle,
+    ConcatBiUniNode(TupleLifecycle<BiTuple<A, B>> nextNodesTupleLifecycle,
             int inputStoreIndexLeftOutTupleList, int inputStoreIndexRightOutTupleList,
             int outputStoreSize) {
         super(nextNodesTupleLifecycle, inputStoreIndexLeftOutTupleList, inputStoreIndexRightOutTupleList, outputStoreSize);
@@ -19,8 +20,8 @@ final class BavetBiBiConcatNode<A, B>
     }
 
     @Override
-    protected BiTuple<A, B> getOutTupleFromRight(BiTuple<A, B> rightTuple) {
-        return new BiTuple<>(rightTuple.factA, rightTuple.factB, outputStoreSize);
+    protected BiTuple<A, B> getOutTupleFromRight(UniTuple<A> rightTuple) {
+        return new BiTuple<>(rightTuple.factA, null, outputStoreSize);
     }
 
     @Override
@@ -30,9 +31,8 @@ final class BavetBiBiConcatNode<A, B>
     }
 
     @Override
-    protected void updateOutTupleFromRight(BiTuple<A, B> rightTuple, BiTuple<A, B> outTuple) {
+    protected void updateOutTupleFromRight(UniTuple<A> rightTuple, BiTuple<A, B> outTuple) {
         outTuple.factA = rightTuple.factA;
-        outTuple.factB = rightTuple.factB;
     }
 
 }
