+++ core/constraint-streams/src/main/java/ai/timefold/solver/constraint/streams/bavet/quad/ConcatQuadBiNode.java
@@ -1,15 +1,16 @@
 package ai.timefold.solver.constraint.streams.bavet.quad;
 
 import ai.timefold.solver.constraint.streams.bavet.common.AbstractConcatNode;
+import ai.timefold.solver.constraint.streams.bavet.common.tuple.BiTuple;
 import ai.timefold.solver.constraint.streams.bavet.common.tuple.QuadTuple;
 import ai.timefold.solver.constraint.streams.bavet.common.tuple.TupleLifecycle;
 
-final class BavetQuadQuadConcatNode<A, B, C, D>
-        extends AbstractConcatNode<QuadTuple<A, B, C, D>, QuadTuple<A, B, C, D>, QuadTuple<A, B, C, D>> {
+final class ConcatQuadBiNode<A, B, C, D>
+        extends AbstractConcatNode<QuadTuple<A, B, C, D>, BiTuple<A, B>, QuadTuple<A, B, C, D>> {
 
-    BavetQuadQuadConcatNode(TupleLifecycle<QuadTuple<A, B, C, D>> nextNodesTupleLifecycle,
+    ConcatQuadBiNode(TupleLifecycle<QuadTuple<A, B, C, D>> nextNodesTupleLifecycle,
             int inputStoreIndexLeftOutTupleList, int inputStoreIndexRightOutTupleList,
-                            int outputStoreSize) {
+            int outputStoreSize) {
         super(nextNodesTupleLifecycle, inputStoreIndexLeftOutTupleList, inputStoreIndexRightOutTupleList, outputStoreSize);
     }
 
@@ -19,8 +20,8 @@ final class BavetQuadQuadConcatNode<A, B, C, D>
     }
 
     @Override
-    protected QuadTuple<A, B, C, D> getOutTupleFromRight(QuadTuple<A, B, C, D> rightTuple) {
-        return new QuadTuple<>(rightTuple.factA, rightTuple.factB, rightTuple.factC, rightTuple.factD, outputStoreSize);
+    protected QuadTuple<A, B, C, D> getOutTupleFromRight(BiTuple<A, B> rightTuple) {
+        return new QuadTuple<>(rightTuple.factA, rightTuple.factB, null, null, outputStoreSize);
     }
 
     @Override
@@ -32,11 +33,9 @@ final class BavetQuadQuadConcatNode<A, B, C, D>
     }
 
     @Override
-    protected void updateOutTupleFromRight(QuadTuple<A, B, C, D> rightTuple, QuadTuple<A, B, C, D> outTuple) {
+    protected void updateOutTupleFromRight(BiTuple<A, B> rightTuple, QuadTuple<A, B, C, D> outTuple) {
         outTuple.factA = rightTuple.factA;
         outTuple.factB = rightTuple.factB;
-        outTuple.factC = rightTuple.factC;
-        outTuple.factD = rightTuple.factD;
     }
 
 }
