+++ core/constraint-streams/src/main/java/ai/timefold/solver/constraint/streams/bavet/quad/ConcatUniQuadNode.java
@@ -1,22 +1,23 @@
 package ai.timefold.solver.constraint.streams.bavet.quad;
 
 import ai.timefold.solver.constraint.streams.bavet.common.AbstractConcatNode;
-import ai.timefold.solver.constraint.streams.bavet.common.tuple.BiTuple;
 import ai.timefold.solver.constraint.streams.bavet.common.tuple.QuadTuple;
 import ai.timefold.solver.constraint.streams.bavet.common.tuple.TupleLifecycle;
+import ai.timefold.solver.constraint.streams.bavet.common.tuple.UniTuple;
 
-final class BavetBiQuadConcatNode<A, B, C, D>
-        extends AbstractConcatNode<BiTuple<A, B>, QuadTuple<A, B, C, D>, QuadTuple<A, B, C, D>> {
+final class ConcatUniQuadNode<A, B, C, D>
+        extends AbstractConcatNode<UniTuple<A>, QuadTuple<A, B, C, D>, QuadTuple<A, B, C, D>> {
 
-    BavetBiQuadConcatNode(TupleLifecycle<QuadTuple<A, B, C, D>> nextNodesTupleLifecycle,
+    ConcatUniQuadNode(TupleLifecycle<QuadTuple<A, B, C, D>> nextNodesTupleLifecycle,
             int inputStoreIndexLeftOutTupleList, int inputStoreIndexRightOutTupleList,
             int outputStoreSize) {
-        super(nextNodesTupleLifecycle, inputStoreIndexLeftOutTupleList, inputStoreIndexRightOutTupleList, outputStoreSize);
+        super(nextNodesTupleLifecycle, inputStoreIndexLeftOutTupleList, inputStoreIndexRightOutTupleList,
+                outputStoreSize);
     }
 
     @Override
-    protected QuadTuple<A, B, C, D> getOutTupleFromLeft(BiTuple<A, B> leftTuple) {
-        return new QuadTuple<>(leftTuple.factA, leftTuple.factB, null, null, outputStoreSize);
+    protected QuadTuple<A, B, C, D> getOutTupleFromLeft(UniTuple<A> leftTuple) {
+        return new QuadTuple<>(leftTuple.factA, null, null, null, outputStoreSize);
     }
 
     @Override
@@ -25,9 +26,8 @@ final class BavetBiQuadConcatNode<A, B, C, D>
     }
 
     @Override
-    protected void updateOutTupleFromLeft(BiTuple<A, B> leftTuple, QuadTuple<A, B, C, D> outTuple) {
+    protected void updateOutTupleFromLeft(UniTuple<A> leftTuple, QuadTuple<A, B, C, D> outTuple) {
         outTuple.factA = leftTuple.factA;
-        outTuple.factB = leftTuple.factB;
     }
 
     @Override
