--- core/constraint-streams/src/main/java/ai/timefold/solver/constraint/streams/bavet/tri/BavetTriBiConcatNode.java
@@ -1,40 +1,0 @@
-package ai.timefold.solver.constraint.streams.bavet.tri;
-
-import ai.timefold.solver.constraint.streams.bavet.common.AbstractConcatNode;
-import ai.timefold.solver.constraint.streams.bavet.common.tuple.BiTuple;
-import ai.timefold.solver.constraint.streams.bavet.common.tuple.TriTuple;
-import ai.timefold.solver.constraint.streams.bavet.common.tuple.TupleLifecycle;
-
-final class BavetTriBiConcatNode<A, B, C>
-        extends AbstractConcatNode<TriTuple<A, B, C>, BiTuple<A, B>, TriTuple<A, B, C>> {
-
-    BavetTriBiConcatNode(TupleLifecycle<TriTuple<A, B, C>> nextNodesTupleLifecycle,
-                         int inputStoreIndexLeftOutTupleList, int inputStoreIndexRightOutTupleList,
-            int outputStoreSize) {
-        super(nextNodesTupleLifecycle, inputStoreIndexLeftOutTupleList, inputStoreIndexRightOutTupleList, outputStoreSize);
-    }
-
-    @Override
-    protected TriTuple<A, B, C> getOutTupleFromLeft(TriTuple<A, B, C> leftTuple) {
-        return new TriTuple<>(leftTuple.factA, leftTuple.factB, leftTuple.factC, outputStoreSize);
-    }
-
-    @Override
-    protected TriTuple<A, B, C> getOutTupleFromRight(BiTuple<A, B> rightTuple) {
-        return new TriTuple<>(rightTuple.factA, rightTuple.factB, null, outputStoreSize);
-    }
-
-    @Override
-    protected void updateOutTupleFromLeft(TriTuple<A, B, C> leftTuple, TriTuple<A, B, C> outTuple) {
-        outTuple.factA = leftTuple.factA;
-        outTuple.factB = leftTuple.factB;
-        outTuple.factC = leftTuple.factC;
-    }
-
-    @Override
-    protected void updateOutTupleFromRight(BiTuple<A, B> rightTuple, TriTuple<A, B, C> outTuple) {
-        outTuple.factA = rightTuple.factA;
-        outTuple.factB = rightTuple.factB;
-    }
-
-}
