+++ core/constraint-streams/src/main/java/ai/timefold/solver/constraint/streams/bavet/tri/ConcatTriBiNode.java
@@ -5,36 +5,36 @@ import ai.timefold.solver.constraint.streams.bavet.common.tuple.BiTuple;
 import ai.timefold.solver.constraint.streams.bavet.common.tuple.TriTuple;
 import ai.timefold.solver.constraint.streams.bavet.common.tuple.TupleLifecycle;
 
-final class BavetBiTriConcatNode<A, B, C>
-        extends AbstractConcatNode<BiTuple<A, B>, TriTuple<A, B, C>, TriTuple<A, B, C>> {
+final class ConcatTriBiNode<A, B, C>
+        extends AbstractConcatNode<TriTuple<A, B, C>, BiTuple<A, B>, TriTuple<A, B, C>> {
 
-    BavetBiTriConcatNode(TupleLifecycle<TriTuple<A, B, C>> nextNodesTupleLifecycle,
+    ConcatTriBiNode(TupleLifecycle<TriTuple<A, B, C>> nextNodesTupleLifecycle,
             int inputStoreIndexLeftOutTupleList, int inputStoreIndexRightOutTupleList,
             int outputStoreSize) {
         super(nextNodesTupleLifecycle, inputStoreIndexLeftOutTupleList, inputStoreIndexRightOutTupleList, outputStoreSize);
     }
 
     @Override
-    protected TriTuple<A, B, C> getOutTupleFromLeft(BiTuple<A, B> leftTuple) {
-        return new TriTuple<>(leftTuple.factA, leftTuple.factB, null, outputStoreSize);
+    protected TriTuple<A, B, C> getOutTupleFromLeft(TriTuple<A, B, C> leftTuple) {
+        return new TriTuple<>(leftTuple.factA, leftTuple.factB, leftTuple.factC, outputStoreSize);
     }
 
     @Override
-    protected TriTuple<A, B, C> getOutTupleFromRight(TriTuple<A, B, C> rightTuple) {
-        return new TriTuple<>(rightTuple.factA, rightTuple.factB, rightTuple.factC, outputStoreSize);
+    protected TriTuple<A, B, C> getOutTupleFromRight(BiTuple<A, B> rightTuple) {
+        return new TriTuple<>(rightTuple.factA, rightTuple.factB, null, outputStoreSize);
     }
 
     @Override
-    protected void updateOutTupleFromLeft(BiTuple<A, B> leftTuple, TriTuple<A, B, C> outTuple) {
+    protected void updateOutTupleFromLeft(TriTuple<A, B, C> leftTuple, TriTuple<A, B, C> outTuple) {
         outTuple.factA = leftTuple.factA;
         outTuple.factB = leftTuple.factB;
+        outTuple.factC = leftTuple.factC;
     }
 
     @Override
-    protected void updateOutTupleFromRight(TriTuple<A, B, C> rightTuple, TriTuple<A, B, C> outTuple) {
+    protected void updateOutTupleFromRight(BiTuple<A, B> rightTuple, TriTuple<A, B, C> outTuple) {
         outTuple.factA = rightTuple.factA;
         outTuple.factB = rightTuple.factB;
-        outTuple.factC = rightTuple.factC;
     }
 
 }
